refactor: 저장소 구현체 분리
