+++ app/src/main/java/woowacourse/shopping/di/auto/AppContainer.kt
@@ -1,11 +1,13 @@
 package woowacourse.shopping.di.auto
 
 import woowacourse.shopping.data.CartRepository
+import woowacourse.shopping.data.CartRepositoryImpl
 import woowacourse.shopping.data.ProductRepository
+import woowacourse.shopping.data.ProductRepositoryImpl
 
 object AppContainer : Container() {
     init {
-        bindSingleton(CartRepository::class) { CartRepository() }
-        bindSingleton(ProductRepository::class) { ProductRepository() }
+        bindSingleton(CartRepository::class) { CartRepositoryImpl() }
+        bindSingleton(ProductRepository::class) { ProductRepositoryImpl() }
     }
 }
