+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -9,7 +9,6 @@ import woowacourse.shopping.model.Product
 class CartViewModel(
     private val cartRepository: CartRepository,
 ) : ViewModel() {
-
     private val _cartProducts: MutableLiveData<List<Product>> =
         MutableLiveData(emptyList())
     val cartProducts: LiveData<List<Product>> get() = _cartProducts
